WS Done

+++ FIFOPager.cpp
@@ -14,7 +14,7 @@ FIFOPager::FIFOPager(int num_frame) : Pager(num_frame) {
 	counter = -1;
 }
 
-Frame* FIFOPager::select_victim_frame(int num_instruction) {
+Frame* FIFOPager::select_victim_frame(unsigned long num_instruction) {
 	counter = (counter + 1) % f_table->GetMaxNum();
 	return f_table->At(counter);
 }
